requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.2
Pillow>=10.0.0
pytesseract>=0.3.10
